chore(frontend): refactor hooks.ts to individual files for each hook
