+++ frontend/src/pages/AuthRedirect.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../utils/hooks';
+import { useAuth } from '../hooks/useAuth';
 import { CircularProgress } from '@mui/material';
 
 export default function AuthRedirect() {
