+++ frontend/src/pages/Profile.tsx
@@ -11,7 +11,7 @@ import {
 } from '@mui/material';
 import { useEffect, useState } from 'react';
 import { getAllLanguages, deleteUserProfile } from '../api/userApi';
-import { useAuth } from '../utils/hooks';
+import { useAuth } from '../hooks/useAuth';
 import { useNavigate } from 'react-router-dom';
 import { Language } from '../@types/language';
 import { useProfile } from '../hooks/useProfile';
