+++ frontend/src/routes/ProtectedRoutes.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
-import { useAuth } from '../utils/hooks';
+import { useAuth } from '../hooks/useAuth';
 import { ProfileProvider } from '../contextProviders/ProfileContext';
 
 export default function ProtectedRoutes() {
